♻️ refactor: return decision context objects from IPlayerActor methods
Rework IPlayerActor interface to return rich decision context objects
(CallTrumpDecisionContext, CardDecisionContext, RelativeCardDecisionContext)
instead of raw values, enabling bots to report predicted points for each
valid decision. Flow predicted points through DecisionRecorder into
decision records and display actual ML scores in GameResultsRenderer.

Also adds models/ to .gitignore for ML model artifacts.

+++ NemesisEuchre.GameEngine/Models/CallTrumpDecisionRecord.cs
@@ -23,5 +23,5 @@ public class CallTrumpDecisionRecord
 
     public CallTrumpDecision ChosenDecision { get; set; }
 
-    public Dictionary<CallTrumpDecision, decimal> DecisionPredictedPoints { get; set; } = [];
+    public Dictionary<CallTrumpDecision, float> DecisionPredictedPoints { get; set; } = [];
 }
