+++ NemesisEuchre.GameEngine/Models/DiscardCardDecisionRecord.cs
@@ -22,5 +22,5 @@ public class DiscardCardDecisionRecord
 
     public required Card ChosenCard { get; set; }
 
-    public Dictionary<Card, decimal> DecisionPredictedPoints { get; set; } = [];
+    public Dictionary<Card, float> DecisionPredictedPoints { get; set; } = [];
 }
